requirements.txt
asgiref==3.4.1
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
cryptography==3.3.2
distro==1.5.0
distro-info==1.0
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.12.0
django-extensions==3.1.5
django-on-heroku==1.1.2
fasteners==0.14.1
future==0.18.2
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
jeepney==0.6.0
keyring==22.2.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
monotonic==1.5
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
paramiko==2.7.2
pbr==5.8.0
pexpect==4.8.0
Pillow==8.1.2
protobuf==3.12.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.21
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
python-dateutil==2.8.1
python-debian==0.1.39
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
testresources==2.0.1
urllib3==1.26.2
wadllib==1.3.5
whitenoise==5.3.0
xdg==5.0.0
